job 삭제 confirm 추가.

+++ src/main/frontend/src/components/QuartzJobList.vue
@@ -163,11 +163,17 @@
       },
 
       deleteJob(row) {
-        JobAdapter.deleteJob(row.id).then((res) => {
-          this.jobList = res;
-          this.$message({
-            message: 'delete success',
-            type: 'success'
+        this.$confirm(`Delete the job?`, 'Delete', {
+          confirmButtonText: 'Yes',
+          cancelButtonText: 'No',
+          type: 'warning'
+        }).then(() => {
+          JobAdapter.deleteJob(row.id).then((res) => {
+            this.jobList = res;
+            this.$message({
+              message: 'delete success',
+              type: 'success'
+            });
           });
         });
       },
